getUserDetail, updateUserDetail. auth for notes' methods

+++ functions/APIs/notes.js
@@ -3,6 +3,7 @@ const {db} = require("../utils/admin");
 exports.getAllNotes = (request, response) => {
     db
         .collection('notes')
+        .where('username', '==', request.user.username)
         .orderBy('createdAt', 'desc')
         .get()
         .then((data) => {
@@ -35,7 +36,8 @@ exports.postOneNote = (request, response) => {
     const newNoteItem = {
         title: request.body.title,
         body: request.body.body,
-        createdAt: new Date().toISOString()
+        createdAt: new Date().toISOString(),
+        username: request.user.username
     }
     db
         .collection('notes')
@@ -58,6 +60,9 @@ exports.deleteNote = (request, response) => {
         .then((doc) => {
             if (!doc.exists) {
                 return response.status(404).json({error: 'Note not found'})
+            }
+            if(doc.data().username !== request.user.username){
+                return response.status(403).json({error:"Unauthorized"})
             }
             return document.delete();
         })
